pyproject.toml
[project]
name = "gem-code"
version = "0.1.0"
description = "A lightweight CLI Agent with interactive chat, tool calling, and skill extensions"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.13.0",
    "aiofiles>=24.1.0,<24.2.0",
    "harbor>=0.1.45",
    "mcp>=1.26.0",
    "openai>=2.21.0",
    "py-landlock>=0.1.1",
    "pygame>=2.6.1",
    "pytest>=9.0.2",
    "python-ulid>=3.1.0",
    "rich>=14.3.3",
    "sentence-transformers>=5.2.3",
    "textual>=0.85.0",
    "textual-dev>=1.8.0",
    "trafilatura>=2.0.0",
]

[project.scripts]
gem-code = "main:main"
gem = "main:main"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
